+++ mpapi-next-gen/src/common/appIdentity.ts
