+++ mpapi-next-gen/src/common/dynamoStream.ts
