+++ mpapi-next-gen/src/lambdas/exportSubscriptionTableV2.ts
@@ -1,7 +1,7 @@
 import 'source-map-support/register';
 import { DynamoDBClient, ExportFormat } from '@aws-sdk/client-dynamodb';
 import { ExportTableToPointInTimeCommand } from '@aws-sdk/client-dynamodb';
-import { plusDays } from '../utils/dates';
+import { plusDays } from '../common/dates';
 
 function prefix_creator(stage: string): string {
 	const yesterday = plusDays(new Date(), -1).toISOString().substring(0, 10);
